% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download_data.R
\name{check_downloads}
\alias{check_downloads}
\title{Check integrity of downloaded FASTQ files}
\usage{
check_downloads(
  sample_info = NULL,
  fastqdir = "results/01_FASTQ_files",
  read_count
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info()}.}

\item{fastqdir}{Path to the directory where .fastq files will be stored.
Default: results/01_FASTQ_files.}

\item{read_count_df}{A 2-column data frame with the number of reads for
each run as reported in SRA, which can be obtained with the
function \code{get_read_count()}.}
}
\value{
A 3-column data frame with the following variables:
\describe{
\item{Run}{Character vector of run acessions.}
\item{Status}{Character vector of file integrity status.
If there is nothing wrong with the file, it will be "OK",
and NA otherwise.}
\item{Issue}{Numeric vector with code to issue in each file, if there is
any. See issue legend below.}
}
}
\description{
This file detects FASTQ files that have not been properly downloaded
and flags them, indicating the problem in a "Issue" column. It is an
alternative to md5sum to check file integrity.
}
\details{
This function looks for and flags problematic downloads.
Run accessions that have been properly downloaded will have "OK" in
the \strong{Status} variable. If there is any issue with the run,
the \strong{Status} variable will be NA, with details on the problem
in the \strong{Issue} column. Issues are labeled numbers from 1 to 3,
which mean:
\itemize{
\item 1: File was not downloaded.
\item 2: For paired-end reads, only forward or reverse file was downloaded.
\item 3: Less reads than the expected. The expected number of reads for each
file is available in SRA, and it can be obtained
with \code{get_read_count()}.
}
}
\examples{
data(sample_info)
fastqdir <- system.file("extdata", package = "bears") 
# Create a fake data frame of read count
read_count <- data.frame(Run = "SRR1039508", Reads = 7097)
# Check download
check_downloads(sample_info, fastqdir, read_count)
}
