% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download_data.R
\name{check_md5}
\alias{check_md5}
\title{Check file integrity with md5sum}
\usage{
check_md5(run_accessions = NULL, fastqdir = "results/01_FASTQ_files")
}
\arguments{
\item{run_accessions}{Character vector of run accessions.}

\item{fastqdir}{Path to the directory where .fastq files will be stored.
Default: results/01_FASTQ_files.}
}
\value{
A data frame with variables \strong{Run} and \strong{Status} with
run accession and integrity status, respectively.
}
\description{
Check file integrity with md5sum
}
\examples{
urls <- c(
    "ftp.sra.ebi.ac.uk/vol1/fastq/SRR926/SRR926397/SRR926397_1.fastq.gz",
    "ftp.sra.ebi.ac.uk/vol1/fastq/SRR926/SRR926397/SRR926397_2.fastq.gz"
)
sample_info <- data.frame(
    BioSample = "SAMN01924555",
    Experiment = "SRX245306",
    Run = "SRR926397",
    BioProject = "PRJNA190191", Instrument = "Illumina HiSeq 2000", 
    Layout = "PAIRED"
)
fastqdir <- tempdir()
d <- download_from_ena(
    sample_info, urls = urls, fastqdir = fastqdir, method = "libcurl"
)

# Check MD5
run_accessions <- sample_info$Run
check_md5(run_accessions, fastqdir)
}
