% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir_structure}
\alias{create_dir_structure}
\title{Create a standard directory for bears}
\usage{
create_dir_structure(rootdir = NULL)
}
\arguments{
\item{rootdir}{Root directory where all subdirectories will be created.
If not specified, a directory named "results_YYYY-MM-DD" will be created,
where Y, M and D represent year, month and day, respectively.}
}
\value{
A list of paths with following elements:
\describe{
\item{fastqdir}{Path to directory where FASTQ files are.}
\item{fastqcdir}{Path to directory where FastQC output is stored.}
\item{filtdir}{Path to directory where filtered FASTQ files are.}
\item{mappingdir}{Path to directory where STAR-derived BAM files are.}
\item{salmonindex}{Path to directory where salmon transcriptome index is.}
\item{salmondir}{Path to directory where salmon output files are.}
\item{kallistoindex}{Path to directory where kallisto transcriptome index is.}
\item{kallistodir}{Path to directory where kallisto output files are.}
\item{fcountsdir}{Path to directory where featureCounts output files are.}
\item{stringtiedir}{Path to directory where StringTie output files are.}
}
}
\description{
Create a standard directory for bears
}
\examples{
rootdir <- tempdir()
dirs <- create_dir_structure(rootdir)
}
