% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download_data.R
\name{create_sample_info}
\alias{create_sample_info}
\title{Search the SRA database and create a data frame of sample metadata}
\usage{
create_sample_info(term, retmax = 5000)
}
\arguments{
\item{term}{Character with the search term,
e.g. "Glycine max[ORGN] AND RNA-seq[STRA]".}

\item{retmax}{Numeric with the maximum number of hits returned
by the search.}
}
\value{
A data frame with the following columns:
\describe{
\item{BioSample}{BioSample accession.}
\item{Experiment}{Experiment accession (SRX*).}
\item{Run}{Run accession (SRR*).}
\item{Tissue}{Tissue from which RNA was extracted.}
\item{Pubmed}{Pubmed ID of articles associated with the project, if any.}
\item{BioProject}{Bioproject accession.}
\item{Instrument}{Sequencing instrument}
\item{Layout}{Library layout (single- or paired-end sequencing).}
\item{Selection_method}{Library selection method.}
\item{SRA_sample}{SRA sample accession (SRS*).}
\item{SRA_study}{SRA study accession (SRP*).}
\item{Treatment}{Treatment for the biosample.}
\item{Cultivar}{Plant cultivar.}
\item{Study_title}{Study title, if any.}
\item{Study_abstract}{Study abstract, if any.}
\item{Date}{Date of release.}
\item{Origin}{Country of origin.}
}
}
\description{
Search the SRA database and create a data frame of sample metadata
}
\examples{
term <- "PRJNA80173[GPRJ]"
df <- create_sample_info(term)
}
