% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download_data.R
\name{download_fastq}
\alias{download_fastq}
\title{Download FASTQ files}
\usage{
download_fastq(
  sample_info,
  fastqdir = "results/01_FASTQ_files",
  sradir = "results/00_SRA_files",
  soliddir = "results/01_SOLiD_dir",
  threads = 6
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{fastqdir}{Path to the directory where .fastq files will be stored.
Default: results/01_FASTQ_files.}

\item{sradir}{Path to the directory where .sra files will be temporarily
stored. Default: results/00_SRA_files.}

\item{soliddir}{Path to the directory where .fastq files for SOLiD data
will be stored. Default: results/01_SOLiD_dir.}
}
\value{
NULL, with .sra files in the directory specified in sradir, and
fastq files in the directories specified in fastqdir and soliddir.
}
\description{
Download FASTQ files
}
\examples{
\donttest{
data(sample_info)
download_fastq(sample_info, threads = 20)
}
}
