% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download_data.R
\name{download_from_ena}
\alias{download_from_ena}
\title{Download FASTQ files from ENA's ftp}
\usage{
download_from_ena(
  sample_info = NULL,
  fastqdir = "results/01_FASTQ_files",
  method = "auto"
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{fastqdir}{Path to the directory where .fastq files will be stored.
Default: results/01_FASTQ_files.}
}
\value{
A data frame as returned by \code{fastq_exists}.
}
\description{
Download FASTQ files from ENA's ftp
}
\examples{
data(sample_info)
fastqdir <- tempdir()
download_from_ena(sample_info, fastqdir)
}
