% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{fastp_is_installed}
\alias{fastp_is_installed}
\title{Check if fastp is installed}
\usage{
fastp_is_installed()
}
\value{
Logical indicating whether fastp is installed or not.
}
\description{
Check if fastp is installed
}
\examples{
fastp_is_installed()
}
