% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{fastqc_is_installed}
\alias{fastqc_is_installed}
\title{Check if FastQC is installed}
\usage{
fastqc_is_installed()
}
\value{
Logical indicating whether FastQC is installed or not.
}
\description{
Check if FastQC is installed
}
\examples{
fastqc_is_installed()
}
