% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fastqc_table}
\alias{fastqc_table}
\title{FastQC summary table}
\format{
A data frame with FastQC summary statistics.
}
\usage{
data(fastqc_table)
}
\description{
This data frame was created with \code{multiqc}. The code can be found in
the script datasets.R
}
\examples{
data(fastqc_table)
}
\keyword{datasets}
