% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_03_quantification_featureCounts.R
\name{featureCounts}
\alias{featureCounts}
\title{Count reads with featureCounts}
\usage{
featureCounts(
  sample_info = NULL,
  mappingdir = "results/04_read_mapping",
  gff_path = NULL,
  fcountsdir = "results/05_quantification/featureCounts",
  threads = 2
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
functions \code{create_sample_info} and \code{infer_strandedness}.}

\item{mappingdir}{Directory where .bam files are stored.}

\item{gff_path}{Path to GFF/GTF file with annotations.}

\item{fcountsdir}{Directory where the matrix of gene-level read counts will
be stored.}

\item{threads}{Number of threads for featureCounts. Default: 2.}
}
\value{
A gene expression matrix with genes in row names and samples in
column names. Two .tsv files with the gene expression matrix and count stats
are saved to \code{fcountsdir}.
}
\description{
Count reads with featureCounts
}
\examples{
data(sample_info)
mappingdir <- system.file("extdata", package="bears")
gff_path <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf",
                        package="bears")
fcountsdir <- file.path(tempdir(), "fcountsdir")
if(subread_is_installed()) {
    counts <- featureCounts(sample_info, mappingdir, gff_path, fcountsdir)
}
}
