% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_03_quantification_featureCounts.R
\name{featureCounts2se}
\alias{featureCounts2se}
\title{Create a SummarizedExperiment object from featureCounts output}
\usage{
featureCounts2se(sample_info = NULL, fc_output = NULL)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
functions \code{create_sample_info}.}

\item{fc_output}{Either a gene expression matrix generated by \code{fcount},
or the path to the .tsv file containing the gene expression matrix as
generated by \code{fcount}.}
}
\value{
A SummarizedExperiment object.
}
\description{
Create a SummarizedExperiment object from featureCounts output
}
\examples{
data(sample_info)
mappingdir <- system.file("extdata", package="bears")
gff_path <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf",
                        package="bears")
fcountsdir <- file.path(tempdir(), "fcountsdir")
if(subread_is_installed()) {
    counts <- featureCounts(sample_info, mappingdir, gff_path, fcountsdir)
    se <- featureCounts2se(sample_info, counts)
}
}
