% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_metadata.R
\name{get_read_count}
\alias{get_read_count}
\title{Get read count information from SRA for each run accession}
\usage{
get_read_count(sample_info, run_accession)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{run_accession}{Character scalar or vector with run accessions
for which read count will be obtained.}
}
\value{
A 2-column data frame with variables \strong{Run} and \strong{Reads}
indicating run accession and number of reads for each sample, respectively.
}
\description{
This function is used in quality checks, so users can check if the files
they have downloaded have the same number of reads as reported in SRA.
If local files have less reads than what is reported in SRA metadata,
probably there was an error during download.
}
\examples{
data(sample_info)
readcount <- get_read_count(sample_info, sample_info$Run)
}
