% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gff2bed}
\alias{gff2bed}
\title{Convert GFF file to BED}
\usage{
gff2bed(gffpath = NULL)
}
\arguments{
\item{gffpath}{Path to .gff file with genome annotation.}
}
\value{
A NULL object.
}
\description{
Convert GFF file to BED
}
\examples{
gffpath <- system.file("extdata", "Gmax_chr15_subset.gff3", package="bears")
gffdir <- tempdir()
file.copy(from = gffpath, to=gffdir)
gff_file <- list.files(gffdir, full.names=TRUE, pattern=".gff3")
gff2bed(gff_file)
}
