% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_02_quantification_kallisto.R
\name{kallisto2se}
\alias{kallisto2se}
\title{Create a SummarizedExperiment object from kallisto output}
\usage{
kallisto2se(
  sample_info = NULL,
  level = "gene",
  kallistodir = "results/05_quantification/kallisto",
  tx2gene = NULL
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
functions \code{create_sample_info}}

\item{level}{Character indicating to which level expression must be
quantified in the SE object. One of "gene" (default), "transcript",
or "both". For "both", the SE object will have two assays named "transcript"
and "gene".}

\item{kallistodir}{Directory where quantification files will be stored.
Default: results/05_quantification/kallisto.}

\item{tx2gene}{Data frame of correspondence between genes and transcripts,
with gene IDs in the first column and transcript IDs in the second column.
Only required if level = 'gene' or 'both'.}
}
\value{
A SummarizedExperiment object with gene/transcript expression
levels and sample metadata.
}
\description{
Create a SummarizedExperiment object from kallisto output
}
\examples{
data(sample_info)
data(tx2gene)
kallistodir <- system.file("extdata", package = "bears")
se_gene <- kallisto2se(sample_info, kallistodir = kallistodir, 
                       tx2gene = tx2gene)
}
