% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_02_quantification_kallisto.R
\name{kallisto_index}
\alias{kallisto_index}
\title{Index the transcriptome for kallisto}
\usage{
kallisto_index(
  kallistoindex = "results/05_quantification/kallisto/idx",
  transcriptome_path = NULL,
  envname = NULL,
  miniconda_path = NULL
)
}
\arguments{
\item{kallistoindex}{Directory where kallisto index file will be stored.
Default: results/05_quantification/kallisto/idx.}

\item{transcriptome_path}{Path to the reference transcriptome FASTA file.}

\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A 2-column data frame with path to index file in the first column
and index build status in the second column, with "OK" if the transcriptome
index was successfully created, and NA otherwise.
}
\description{
Index the transcriptome for kallisto
}
\examples{
kallistoindex <- file.path(tempdir(), "idx")
transcriptome_path <- system.file(
     "extdata", "Hsapiens_GRCh37.75_subset_transcripts.fa", package="bears"
)
if(kallisto_is_installed()) {
    kallisto_index(kallistoindex, transcriptome_path)
}
}
