% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{kallisto_is_installed}
\alias{kallisto_is_installed}
\title{Check if kallisto is installed}
\usage{
kallisto_is_installed()
}
\value{
Logical indicating whether kallisto is installed or not
}
\description{
Check if kallisto is installed
}
\examples{
kallisto_is_installed()
}
