% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_read_mapping.R
\name{mapping_pass}
\alias{mapping_pass}
\title{Keep only samples that passed minimum requirements in STAR alignment}
\usage{
mapping_pass(mapping_qc = NULL, sample_info = NULL)
}
\arguments{
\item{mapping_qc}{Data frame with read mapping statistics, as generated by
\code{multiqc()}.}

\item{sample_info}{Data frame with sample metadata as generated by
\code{create_sample_info}.}
}
\value{
Data frame with metadata of samples that passed alignment QC.
}
\description{
Keep only samples that passed minimum requirements in STAR alignment
}
\details{
Samples are excluded if i. more than 50 percent of the reads
fail to map, or ii. more than 40 percent of the reads fail to uniquely map.
}
\examples{
data(sample_info)
dir <- system.file("extdata", package="bears")
out <- tempdir()
if(multiqc_is_installed()) {
    mapping_qc <- multiqc(dir, out, runon="star")
    mapping_passed <- mapping_pass(mapping_qc, sample_info)
}
}
