% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_qc}
\alias{mapping_qc}
\title{Summary statistics of STAR mapping QC}
\format{
A data frame with read mapping QC summary statistics.
}
\usage{
data(mapping_qc)
}
\description{
This data frame was created with \code{multiqc}. The code can be found in
the script datasets.R
}
\examples{
data(mapping_qc)
}
\keyword{datasets}
