% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{multiqc}
\alias{multiqc}
\title{Run MultiQC to get a summary of QC results}
\usage{
multiqc(dir = "results/02_FastQC_dir", outdir = "results/multiqc/fastqc")
}
\arguments{
\item{dir}{Directory where MultiQC should be run.
Default: results/02_FastQC_dir (for FastQC results).}

\item{outdir}{Path the output directory.
Default: results/multiqc/fastqc.}
}
\value{
A data frame of summary statistics for FastQC.
}
\description{
Run MultiQC to get a summary of QC results
}
