% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{multiqc_is_installed}
\alias{multiqc_is_installed}
\title{Check if multiqc is installed}
\usage{
multiqc_is_installed()
}
\value{
Logical indicating whether MultiQC is installed or not.
}
\description{
Check if multiqc is installed
}
\examples{
multiqc_is_installed()
}
