% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filtering.R
\name{remove_rrna}
\alias{remove_rrna}
\title{Remove rRNA sequences from .fastq files with SortMeRna}
\usage{
remove_rrna(
  sample_info,
  fastqdir = "results/01_FASTQ_files",
  filtdir = "results/03_filtered_FASTQ",
  threads = 8
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{fastqdir}{Path to the directory where .fastq files are stored.}

\item{filtdir}{Path to the directory where filtered reads will be stored.
Default: results/03_filtered_FASTQ}

\item{threads}{Number of threads for SortMeRna. Default: 8.}
}
\description{
Remove rRNA sequences from .fastq files with SortMeRna
}
