% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{rseqc_is_installed}
\alias{rseqc_is_installed}
\title{Check if RSeQC is installed}
\usage{
rseqc_is_installed()
}
\value{
Logical indicating whether RSeQC is instaled or not
}
\description{
Check if RSeQC is installed
}
\examples{
rseqc_is_installed()
}
