% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_01_quantification_salmon.R
\name{salmon2se}
\alias{salmon2se}
\title{Create a SummarizedExperiment object from salmon output}
\usage{
salmon2se(
  sample_info = NULL,
  level = "gene",
  salmondir = "results/05_quantification/salmon",
  tx2gene = NULL,
  countsFromAbundance = "lengthScaledTPM"
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
functions \code{create_sample_info}}

\item{level}{Character indicating to which level expression must be
quantified in the SE object. One of "gene" (default), "transcript",
or "both". For "both", the SE object will have two assays named "transcript"
and "gene".}

\item{salmondir}{Directory where quantification files will be stored.
Default: results/05_quantification/salmon.}

\item{tx2gene}{Data frame of correspondence between genes and transcripts,
with gene IDs in the first column and transcript IDs in the second column.
Only required if level = 'gene' or 'both'.}

\item{countsFromAbundance}{Exactly as in \code{tximport::tximport},
whether ot not to generate estimated counts using abundance estimates.
One of "lengthScaledTPM" (default), "scaledTPM", "no", or "dtuScaledTPM".
See ?tximport for details.}
}
\value{
A SummarizedExperiment object with gene/transcript expression
levels and sample metadata.
}
\description{
Create a SummarizedExperiment object from salmon output
}
\examples{
data(sample_info)
data(tx2gene)
salmondir <- system.file("extdata", package="bears")
se_gene <- salmon2se(sample_info, salmondir = salmondir, tx2gene = tx2gene)
}
