% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{salmon_is_installed}
\alias{salmon_is_installed}
\title{Check if salmon is installed}
\usage{
salmon_is_installed()
}
\value{
Logical indicating whether salmon is instaled or not
}
\description{
Check if salmon is installed
}
\examples{
salmon_is_installed()
}
