% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_01_quantification_salmon.R
\name{salmon_quantify}
\alias{salmon_quantify}
\title{Quantify expression with salmon}
\usage{
salmon_quantify(
  sample_info = NULL,
  fastqc_table = NULL,
  filtdir = "results/03_filtered_FASTQ",
  salmonindex = "results/05_quantification/salmon/idx",
  salmondir = "results/05_quantification/salmon",
  threads = NULL,
  envname = NULL,
  miniconda_path = NULL
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
functions \code{create_sample_info} and \code{infer_strandedness}.
The column "Orientation", added by \code{infer_strandedness}, is mandatory.}

\item{fastqc_table}{Data frame of summary statistics for FastQC as returned
by \code{multiqc()}.}

\item{filtdir}{Path to the directory where filtered reads are stored.
Default: results/03_filtered_FASTQ.}

\item{salmonindex}{Directory where the transcriptome index is stored.
Default: results/05_quantification/salmon/idx.}

\item{salmondir}{Directory where quantification files will be stored.
Default: results/05_quantification/salmon.}

\item{threads}{Number of threads for salmon quant.}

\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A NULL object.
}
\description{
Quantify expression with salmon
}
\examples{
data(sample_info)
data(fastqc_table)
filtdir <- system.file("extdata", package = "bears")
salmonindex <- tempdir()
salmondir <- tempdir()
transcriptome_path <- system.file(
     "extdata", "Hsapiens_GRCh37.75_subset_transcripts.fa", package="bears"
)
if(salmon_is_installed()) {
    salmon_index(salmonindex, transcriptome_path)
    salmon_quantify(sample_info, fastqc_table, filtdir, 
                    salmonindex, salmondir)
}

}
