% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{shrimp_is_installed}
\alias{shrimp_is_installed}
\title{Check if SHRiMP is installed}
\usage{
shrimp_is_installed()
}
\value{
Logical indicating whether SHRiMP is instaled or not
}
\description{
Check if SHRiMP is installed
}
\examples{
shrimp_is_installed()
}
