% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_read_mapping.R
\name{solid_align}
\alias{solid_align}
\title{Align SOLiD reads with gmapper-cs}
\usage{
solid_align(
  sample_info = NULL,
  genome_path = NULL,
  soliddir = "results/01_SOLiD_dir",
  mappingdir = "results/04_read_mapping",
  indexdir = "results/04_read_mapping/shrimpIndex",
  threads = 2,
  envname = NULL,
  miniconda_path = NULL
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata.}

\item{genome_path}{Path to the FASTA file representing the genome.}

\item{soliddir}{Path to the directory where SOLiD reads are stored.}

\item{mappingdir}{Path to the directory where .bam files will be stored.}

\item{indexdir}{Path to the directory where genome index will be stored.}

\item{threads}{Numeric representing the number of threads to use.
Default = 2.}

\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A NULL object.
}
\description{
Align SOLiD reads with gmapper-cs
}
\examples{
\donttest{
data(sample_info)
genome_path <- system.file("extdata", "Gmax_chr15_subset.fa", package="bears")
soliddir <- system.file("extdata", package="bears")
mappingdir <- tempdir()
indexdir <- tempdir()
if(shrimp_is_installed()) {
solid_align(sample_info[1, ], genome_path, soliddir, mappingdir,
            indexdir)
}
}
}
