% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{sortmerna_is_installed}
\alias{sortmerna_is_installed}
\title{Check if SortMeRNA is installed}
\usage{
sortmerna_is_installed()
}
\value{
Logical indicating whether SortMeRNA is installed or not.
}
\description{
Check if SortMeRNA is installed
}
\examples{
sortmerna_is_installed()
}
