% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{sratoolkit_is_installed}
\alias{sratoolkit_is_installed}
\title{Check if SRAToolkit is installed}
\usage{
sratoolkit_is_installed()
}
\value{
Logical indicating whether SRAToolkit is installed or not.
}
\description{
Check if SRAToolkit is installed
}
\examples{
sratoolkit_is_installed()
}
