% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_read_mapping.R
\name{star_genome_index}
\alias{star_genome_index}
\title{Index genome for STAR alignment}
\usage{
star_genome_index(
  genome_path = NULL,
  gff_path = NULL,
  mappingdir = "results/04_read_mapping",
  indexdir = "results/04_read_mapping/genomeIndex",
  threads = 10
)
}
\arguments{
\item{genome_path}{Path to the genome .fasta file.}

\item{gff_path}{Path to the .gff/.gtf file with annotations.}

\item{mappingdir}{Path to the directory where read mapping files (.bam) will
be stored.}

\item{indexdir}{Directory where the STAR genome index files will be stored.
Default: results/04_read_mapping/genomeIndex.}

\item{threads}{Number of threads for STAR aligner.}
}
\description{
Index genome for STAR alignment
}
