% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{star_is_installed}
\alias{star_is_installed}
\title{Check if STAR is installed}
\usage{
star_is_installed()
}
\value{
Logical indicating whether STAR is installed or not.
}
\description{
Check if STAR is installed
}
\examples{
star_is_installed()
}
