% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_04_assembly_quantification_stringtie.R
\name{stringtie_assemble}
\alias{stringtie_assemble}
\title{Assemble transcripts with StringTie}
\usage{
stringtie_assemble(
  sample_info = NULL,
  fastqc_table = NULL,
  mappingdir = "results/04_read_mapping",
  gff_path = NULL,
  stringtiedir = "results/05_quantification/stringtie",
  threads = 2,
  envname = NULL,
  miniconda_path = NULL
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
functions \code{create_sample_info} and \code{infer_strandedness}.
The function \code{infer_strandedness} adds a column named "Orientation"
with library strandedness information, which is mandatory.}

\item{fastqc_table}{Data frame of summary statistics for FastQC as returned
by \code{multiqc()}.}

\item{mappingdir}{Directory where .bam files are stored.}

\item{gff_path}{Path to GFF/GTF file with annotations.}

\item{stringtiedir}{Directory where StringTie output files will be stored.}

\item{threads}{Number of threads to use. Default: 2.}

\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A 2-column data frame with samples in the first column and status
in the second column, with "OK" if transcripts were assembled, and
NA otherwise.
}
\description{
Assemble transcripts with StringTie
}
\examples{
data(sample_info)
data(fastqc_table)
mappingdir <- system.file("extdata", package="bears")
gff_path <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf",
                        package="bears")
stringtiedir <- tempdir()
if(stringtie_is_installed()) {
    a <- stringtie_assemble(sample_info, fastqc_table, mappingdir, 
                            gff_path, stringtiedir)
}
}
