% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{stringtie_is_installed}
\alias{stringtie_is_installed}
\title{Check if StringTie is installed}
\usage{
stringtie_is_installed()
}
\value{
Logical indicating whether StringTie is installed or not
}
\description{
Check if StringTie is installed
}
\examples{
stringtie_is_installed()
}
