% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{subread_is_installed}
\alias{subread_is_installed}
\title{Check if subread is installed}
\usage{
subread_is_installed()
}
\value{
Logical indicating whether subread is installed or not
}
\description{
Check if subread is installed
}
\examples{
subread_is_installed()
}
