% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_qc.R
\name{summary_stats_fastp}
\alias{summary_stats_fastp}
\title{Get read quality summary statistics from fastp}
\usage{
summary_stats_fastp(fastp_qcdir = "results/QC_dir/fastp_stats")
}
\arguments{
\item{fastp_qcdir}{Character with path to the directory where .json files
from fastp are stored. Default: results/QC_dir/fastp_stats.}
}
\value{
A data frame of fastp summary stats for each sample with the
following variables:
\itemize{
\item Sample
\item sequencing
\item before_nreads
\item before_nbases
\item before_q20bases
\item before_q30bases
\item before_q20rate
\item before_q30rate
\item before_GCcontent
\item before_meanlength
\item after_nreads
\item after_nbases
\item after_q20bases
\item after_q30bases
\item after_q20rate
\item after_q30rate
\item after_GCcontent
\item after_meanlength
\item filter_n_passed
\item filter_n_lowquality
\item filter_n_too_many_N
\item filter_n_tooshort
\item filter_n_toolong
\item duplication_rate
}
}
\description{
Get read quality summary statistics from fastp
}
\examples{
fastp_qcdir <- system.file("extdata", package = "bears")
fastp_stats <- summary_stats_fastp(fastp_qcdir)
}
