% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_qc.R
\name{summary_stats_salmon}
\alias{summary_stats_salmon}
\title{Get salmon mapping rate for each BioSample}
\usage{
summary_stats_salmon(salmondir, biosamples)
}
\arguments{
\item{salmondir}{Path to the directory where salmon
output files are.}

\item{biosamples}{A character vector of BioSample IDs that were
quantified with salmon.}
}
\value{
A 2-column data frame with the following variables:
\describe{
\item{BioSample}{Character, BioSample IDs.}
\item{Mapping_rate}{Numeric, mapping rate. Values range from 0 to 100.}
}
}
\description{
Get salmon mapping rate for each BioSample
}
\examples{
salmondir <- system.file("extdata", package = "bears")
biosamples <- "SAMN02422669"
summary_stats_salmon(salmondir, biosamples)
}
