% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_qc.R
\name{summary_stats_star}
\alias{summary_stats_star}
\title{Get mapping summary statistics from STAR}
\usage{
summary_stats_star(star_dir = "results/03_read_mapping")
}
\arguments{
\item{star_dir}{Directory where STAR .log files are stored.
Default: results/03_read_mapping.}
}
\value{
A data frame with STAR summary stats for each sample containing the
following variables (all numeric, except \strong{Sample}):
\itemize{
\item Sample
\item total_reads
\item avg_input_read_length
\item uniquely_mapped
\item uniquely_mapped_percent
\item avg_mapped_read_length
\item num_splices
\item num_annotated_splices
\item num_GTAG_splices
\item num_GCAG_splices
\item num_ATAC_splices
\item num_noncanonical_splices
\item mismatch_rate
\item deletion_rate
\item deletion_length
\item insertion_rate
\item insertion_length
\item multimapped
\item multimapped_percent
\item multimapped_toomany
\item multimapped_toomany_percent
\item unmapped_mismatches_percent
\item unmapped_tooshort_percent
\item unmapped_other_percent
\item unmapped_mismatches
\item unmapped_tooshort
\item unmapped_other
}
}
\description{
Get mapping summary statistics from STAR
}
\examples{
star_dir <- system.file("extdata", package = "bears")
qc_table <- summary_stats_star(star_dir)
}
