% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{taco_is_installed}
\alias{taco_is_installed}
\title{Check if TACO is installed}
\usage{
taco_is_installed()
}
\value{
Logical indicating whether TACO is installed or not
}
\description{
Check if TACO is installed
}
\examples{
taco_is_installed()
}
