% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_04_assembly_quantification_stringtie.R
\name{taco_merge}
\alias{taco_merge}
\title{Merge assembled transcripts with TACO}
\usage{
taco_merge(
  sample_info = NULL,
  stringtiedir = "results/05_quantification/stringtie",
  threads = 1
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{stringtiedir}{Directory where StringTie output files will be stored.}

\item{threads}{Number of threads to use. Default: 1.}
}
\value{
A 2-column data frame with the TACO run status, with "OK" if the
program merged transcripts successfully, and NA otherwise.
}
\description{
This function merges GFF files from StringTie assembly into a single GFF.
}
\details{
The GFF files for each BioSample are first merged into
BioProject-level GFF files. Then, these Bioproject-level GFF files are
merged into a single file that represents the whole set.
}
\examples{
data(sample_info)
data(fastqc_table)
mappingdir <- system.file("extdata", package="bears")
gff_path <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf",
                        package="bears")
stringtiedir <- tempdir()
if(stringtie_is_installed()) {
    a <- stringtie_assemble(sample_info, fastqc_table, mappingdir, 
                            gff_path, stringtiedir)
}
if(taco_is_installed()) {
    taco_merge(sample_info, stringtiedir)
}
}
