% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate_strandedness}
\alias{translate_strandedness}
\title{Translate library orientation terminology for each program}
\usage{
translate_strandedness(orientation = NULL, layout = NULL)
}
\arguments{
\item{orientation}{Library orientation, available in
the column "Orientation" from the output of \code{infer_strandedness}.}

\item{layout}{Library layout, available in the column "Layout" from the
output of \code{create_sample_info}.}
}
\value{
A list with the following elements:
\describe{
\item{salmon}{salmon library information.}
\item{fcounts}{featureCounts library information.}
\item{kallisto}{kallisto library information.}
\item{stringtie}{StringTie library information.}
}
}
\description{
Translate library orientation terminology for each program
}
\examples{
data(sample_info)
orientation <- sample_info$Orientation
layout <- sample_info$Layout
strandedness <- translate_strandedness(orientation, layout)
}
