% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{trimmomatic_is_installed}
\alias{trimmomatic_is_installed}
\title{Check if Trimmomatic is installed}
\usage{
trimmomatic_is_installed()
}
\value{
Logical indicating whether Trimmomatic is installed or not.
}
\description{
Check if Trimmomatic is installed
}
\examples{
trimmomatic_is_installed()
}
