% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tx2gene}
\alias{tx2gene}
\title{Transcript to gene mapping}
\format{
A data frame with transcript IDs in the first column and gene IDs
in the second column.
}
\usage{
data(tx2gene)
}
\description{
The code used to create this data frame can be found in
the script datasets.R
}
\examples{
data(tx2gene)
}
\keyword{datasets}
