% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download_data.R
\name{download_fastq}
\alias{download_fastq}
\title{Download FASTQ files}
\usage{
download_fastq(sample_info, fastqdir = "results/01_FASTQ_files", threads = 1)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{fastqdir}{Path to the directory where .fastq files will be stored.
Default: results/01_FASTQ_files.}

\item{threads}{Number of threads to use. Default: 1.}
}
\value{
A 2-column data frame with run accession in the first column
and status in the second column. If file is present, the status "OK"
is returned, otherwise NA is returned.
}
\description{
Download FASTQ files
}
\examples{
\donttest{
data(sample_info)
fastqdir <- tempdir()
if(sratoolkit_is_installed()) {
    download_fastq(sample_info, fastqdir)
}
}
}
