% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download_data.R
\name{get_url_ena}
\alias{get_url_ena}
\title{Get URL for each file in the ENA's FTP repository}
\usage{
get_url_ena(sample_info = NULL, link_from = "api")
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{link_from}{Method to extract the URL to each FASTQ file in the ENA's
ftp repository. One of 'api' or 'iterative'. Default: 'api'.}
}
\value{
A character vector with the URL for each accession.
}
\description{
Get URL for each file in the ENA's FTP repository
}
\examples{
data(sample_info)
get_url_ena(sample_info)
}
