% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gff2bed}
\alias{gff2bed}
\title{Convert GFF file to BED}
\usage{
gff2bed(gffpath = NULL)
}
\arguments{
\item{gffpath}{Path to .gff file with genome annotation.}
}
\value{
Path to output .bed file
}
\description{
Convert GFF file to BED
}
\examples{
gff_path <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf", 
                         package="bears")
gffdir <- tempdir()
file.copy(from = gff_path, to=gffdir)
gff_file <- list.files(gffdir, full.names=TRUE, pattern=".gtf")
gff2bed(gff_file)
}
