% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_strandedness}
\alias{infer_strandedness}
\title{Infer library strandedness}
\usage{
infer_strandedness(
  mapping_passed = NULL,
  bedpath = NULL,
  mappingdir = "results/04_read_mapping"
)
}
\arguments{
\item{mapping_passed}{Metadata of samples that passed mapping QC.
This can be obtained with \code{mapping_pass}.}

\item{bedpath}{Path to BED file. GFF files can be converted to BED with
\code{gff2bed}.}

\item{mappingdir}{Directory where .bam files are stored.}
}
\value{
A data frame with sample metadata as in mapping_passed, but with
an additional column named 'Orientation' containing library strandedness
for each BioProject.
}
\description{
Infer library strandedness
}
\examples{
data(sample_info)
mapping_passed <- sample_info
bedpath <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.bed", 
                        package="bears")
mappingdir <- system.file("extdata", package="bears")
if(rseqc_is_installed()) {
    strandedness <- infer_strandedness(mapping_passed, bedpath, mappingdir)
}

}
