% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_qc.R
\name{multiqc}
\alias{multiqc}
\title{Run MultiQC to get a summary of QC results}
\usage{
multiqc(
  dir = "results/02_FastQC_dir",
  outdir = "results/multiqc/fastqc",
  runon = "fastqc"
)
}
\arguments{
\item{dir}{Directory where MultiQC should be run.
Default: results/02_FastQC_dir (for FastQC results).}

\item{outdir}{Path the output directory.
Default: results/multiqc/fastqc.}

\item{runon}{Type of QC report to generate. One of "fastqc" or "star".}
}
\value{
A data frame with QC summary for each sample.
}
\description{
Run MultiQC to get a summary of QC results
}
\examples{
data(sample_info)
fq <- system.file("extdata", package="bears")
out <- tempdir()
if(multiqc_is_installed()) {
    fastqc_table <- multiqc(fq, out)
}
}
