% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_read_filtering.R
\name{remove_rrna}
\alias{remove_rrna}
\title{Remove rRNA sequences from .fastq files with SortMeRNA}
\usage{
remove_rrna(
  sample_info,
  fastqdir = "results/01_FASTQ_files",
  filtdir = "results/03_filtered_FASTQ",
  rrna_db_dir = NULL,
  threads = 2,
  envname = NULL,
  miniconda_path = NULL
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{fastqdir}{Path to the directory where .fastq files are stored.}

\item{filtdir}{Path to the directory where filtered reads will be stored.
Default: results/03_filtered_FASTQ.}

\item{rrna_db_dir}{Path to directory containing reference rRNA database,
which must be stored as FASTA files.}

\item{threads}{Number of threads for SortMeRna. Default: 2.}

\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A 2-column data frame with run accessions in the first column
and SortMeRNA running status in the second column, with "OK" if SortMeRNA
ran successfully for each file and NA otherwise.
}
\description{
Remove rRNA sequences from .fastq files with SortMeRNA
}
\examples{
data(sample_info)
fastqdir <- system.file("extdata", package="bears")
filtdir <- tempdir()
rrna_db_dir <- tempdir()
rrna_file <- system.file("extdata", "bac_16s_subset.fa", package="bears")
file.copy(from = rrna_file, to = rrna_db_dir)
if(sortmerna_is_installed()) {
    remove_rrna(sample_info, fastqdir, filtdir, rrna_db_dir)
}
}
