% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_qc.R
\name{run_fastqc}
\alias{run_fastqc}
\title{Run FastQC on .fastq files}
\usage{
run_fastqc(
  sample_info,
  fastqdir = "results/01_FASTQ_files",
  fastqcdir = "results/02_FastQC_dir",
  envname = NULL,
  miniconda_path = NULL
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{fastqdir}{Path to the directory where .fastq files will be stored.
Default: results/01_FASTQ_files.}

\item{fastqcdir}{Path to the directory where FastQC output will be stored.
Default: results/02_FastQC_dir}

\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A NULL object.
}
\description{
Run FastQC on .fastq files
}
\examples{
data(sample_info)
fq <- system.file("extdata", package="bear")
fqc <- tempdir()
if(fastqc_is_installed()) {
    run_fastqc(sample_info[1, ], fastqdir = fq, fastqcdir = fqc)
}
}
