% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_01_quantification_salmon.R
\name{salmon_index}
\alias{salmon_index}
\title{Index the transcriptome for salmon}
\usage{
salmon_index(
  salmonindex = "results/05_quantification/salmon/idx",
  transcriptome_path = NULL,
  klen = 31,
  envname = NULL,
  miniconda_path = NULL
)
}
\arguments{
\item{salmonindex}{Directory where the transcriptome index will be stored.
Default: results/05_quantification/salmon/idx.}

\item{transcriptome_path}{Path to the reference transcriptome FASTA file.}

\item{klen}{K-mer length. Default: 31.}

\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A 2-column data frame with path to index in the first column and
index build status in the second column, with "OK" if the transcriptome
index was successfully created, and NA otherwise.
}
\description{
Index the transcriptome for salmon
}
\examples{
salmonindex <- tempdir()
transcriptome_path <- system.file(
     "extdata", "Hsapiens_GRCh37.75_subset_transcripts.fa", package="bears"
)
if(salmon_is_installed()) {
    salmon_index(salmonindex, transcriptome_path)
}
}
