% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_01_quantification_salmon.R
\name{salmon_quantify}
\alias{salmon_quantify}
\title{Quantify expression with salmon}
\usage{
salmon_quantify(
  sample_info = NULL,
  filtdir = "results/03_filtered_FASTQ",
  salmonindex = "results/05_quantification/salmon/idx",
  salmondir = "results/05_quantification/salmon",
  threads = NULL
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
functions \code{create_sample_info} and \code{infer_strandedness}.
The function \code{infer_strandedness} adds a column named "Orientation"
with library strandedness information. If this column is not present in
sample_info, salmon will automatically infer the library strandedness, but
it will take longer to run.}

\item{filtdir}{Path to the directory where filtered reads are stored.
Default: results/03_filtered_FASTQ.}

\item{salmonindex}{Directory where the transcriptome index is stored.
Default: results/05_quantification/salmon/idx.}

\item{salmondir}{Directory where quantification files will be stored.
Default: results/05_quantification/salmon.}

\item{threads}{Number of threads for salmon quant.}
}
\value{
A 2-column data frame with BioSample IDs in the first column and
salmon quantification status in the second column, with "OK" if salmon
sucessfully quantified expression for a given BioSample, and NA otherwise.
}
\description{
Quantify expression with salmon
}
\examples{
data(sample_info)
filtdir <- system.file("extdata", package = "bears")
salmonindex <- tempdir()
salmondir <- tempdir()
transcriptome_path <- system.file(
     "extdata", "Hsapiens_GRCh37.75_subset_transcripts.fa", package="bears"
)
if(salmon_is_installed()) {
    salmon_index(salmonindex, transcriptome_path)
    salmon_quantify(sample_info, filtdir, salmonindex, salmondir)
}
}
