% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_info}
\alias{sample_info}
\title{Sample metadata}
\format{
A data frame with sample metadata created with
\code{create_sample_info}.
}
\usage{
data(sample_info)
}
\description{
This data frame was created with \code{create_sample_info}, and it contains
metadata associated with the BioSamples SAMN05300278 (single-end),
SAMD00117551 (paired-end), and SAMN02688407 (SOLiD).
}
\examples{
data(sample_info)
}
\keyword{datasets}
