% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_read_mapping.R
\name{solid_align}
\alias{solid_align}
\title{Align SOLiD reads with gmapper-cs}
\usage{
solid_align(
  sample_info = NULL,
  genome_path = NULL,
  soliddir = "results/01_SOLiD_dir",
  mappingdir = "results/04_read_mapping",
  indexdir = "results/04_read_mapping/shrimpIndex",
  threads = 20
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata.}

\item{genome_path}{Path to the FASTA file representing the genome.}

\item{soliddir}{Path to the directory where SOLiD reads are stored.}

\item{mappingdir}{Path to the directory where .bam files will be stored.}

\item{indexdir}{Path to the directory where genome index will be stored.}

\item{threads}{Numeric representing the number of threads to use.
Default = 20.}
}
\description{
Align SOLiD reads with gmapper-cs
}
