% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_read_mapping.R
\name{star_align}
\alias{star_align}
\title{Align reads to a reference genome using STAR}
\usage{
star_align(
  sample_info = NULL,
  filtdir = "results/03_filtered_FASTQ",
  fastqc_table = NULL,
  mappingdir = "results/04_read_mapping",
  indexdir = "results/04_read_mapping/genomeIndex",
  gff_path = NULL,
  threads = 2,
  envname = NULL,
  miniconda_path = NULL
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{filtdir}{Path to the directory where filtered reads will be stored.
Default: results/03_filtered_FASTQ.}

\item{fastqc_table}{Data frame of summary statistics for FastQC as returned
by \code{multiqc()}.}

\item{mappingdir}{Path to the directory where read mapping files (.bam) will
be stored.}

\item{indexdir}{Directory where the STAR genome index files will be stored.
Default: results/04_read_mapping/genomeIndex.}

\item{gff_path}{Path to the .gff/.gtf file with annotations.}

\item{threads}{Number of threads for STAR aligner. Default: 2.}

\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A 2-column data frame with BioSample IDs in the first column
and STAR running status in the second column, with "OK" if reads
were mapped (.bam files were created) and NA if STAR failed to map reads.
}
\description{
Align reads to a reference genome using STAR
}
\examples{
\donttest{
data(sample_info)
data(fastqc_table)
genome_path <- system.file("extdata", "Hsapiens_GRCh37.75_subset.fa", 
                            package="bears")
gff_path <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf", 
                         package="bears")
mappingdir <- tempdir()
indexdir <- tempdir()
threads <- 2
filtdir <- system.file("extdata", package="bears")
if(star_is_installed()) {
    star_genome_index(genome_path, gff_path, mapping_dir, indexdir)
    star_align(sample_info, filtdir, fastqc_table, mappingdir, 
               indexdir, gff_path, threads)
}
}
}
