% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_read_mapping.R
\name{star_genome_index}
\alias{star_genome_index}
\title{Index genome for STAR alignment}
\usage{
star_genome_index(
  genome_path = NULL,
  gff_path = NULL,
  mappingdir = "results/04_read_mapping",
  threads = 2
)
}
\arguments{
\item{genome_path}{Path to the genome .fasta file.}

\item{gff_path}{Path to the .gff/.gtf file with annotations.}

\item{mappingdir}{Path to the directory where read mapping files (.bam) will
be stored.}

\item{threads}{Number of threads for STAR aligner. Default: 2.}
}
\value{
A 2-column data frame with path to index in the first column and
index build status in the second column, with "OK" if the index was
built successfully and NA otherwise.
}
\description{
Index genome for STAR alignment
}
\examples{
\donttest{
genome_path <- system.file("extdata", "Hsapiens_GRCh37.75_subset.fa", 
                            package="bears")
gff_path <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf", 
                         package="bears")
mappingdir <- tempdir()
if(star_is_installed()) {
    star_genome_index(genome_path, gff_path, mapping_dir)
}
}
}
