% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_read_filtering.R
\name{trim_reads}
\alias{trim_reads}
\title{Run Trimmomatic on reads that failed on FastQC check}
\usage{
trim_reads(
  sample_info = NULL,
  fastqc_table = NULL,
  fastqdir = "results/01_FASTQ_files",
  filtdir = "results/03_filtered_FASTQ",
  adapterdir = "results/03_filtered_FASTQ/adapters"
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{fastqc_table}{Data frame of summary statistics for FastQC as returned
by \code{multiqc()}.}

\item{fastqdir}{Path to the directory where .fastq files will be stored.
Default: results/01_FASTQ_files.}

\item{filtdir}{Path to the directory where filtered .fastq files will
be stored. Default: results/03_filtered_FASTQ.}
}
\description{
Run Trimmomatic on reads that failed on FastQC check
}
