% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_read_filtering.R
\name{trim_reads}
\alias{trim_reads}
\title{Run Trimmomatic on reads that failed on FastQC check}
\usage{
trim_reads(
  sample_info = NULL,
  fastqc_table = NULL,
  fastqdir = "results/01_FASTQ_files",
  filtdir = "results/03_filtered_FASTQ"
)
}
\arguments{
\item{sample_info}{Data frame of sample metadata created with the
function \code{create_sample_info}.}

\item{fastqc_table}{Data frame of summary statistics for FastQC as returned
by \code{multiqc()}.}

\item{fastqdir}{Path to the directory where .fastq files will be stored.
Default: results/01_FASTQ_files.}

\item{filtdir}{Path to the directory where filtered .fastq files will
be temporarily stored. After trimming, filtered reads are moved back to
fastqdir. Default: results/03_filtered_FASTQ.}
}
\value{
A 2-column data frame with run accession in the first column
and Trimmomatic run status in the second column, with "OK" if it ran
successfully and NA if a file could not be run.
}
\description{
Run Trimmomatic on reads that failed on FastQC check
}
\examples{
data(sample_info)
data(fastqc_table)
fastqdir <- tempdir()
file.copy(system.file("extdata", "SRR1039508_1.fastq.gz", package = "bears"),
          fastqdir)
file.copy(system.file("extdata", "SRR1039508_2.fastq.gz", package = "bears"),
          fastqdir)
dir.create(paste0(fastqdir, "/filtdir"))
filtdir <- paste0(fastqdir, "/filtdir")
if(trimmomatic_is_installed()) {
    trim_reads(sample_info, fastqc_table, fastqdir, filtdir)
}
}
